requirements.txt
langchain-core==0.3.67
langchain_openai==0.3.27
langgraph==0.5.0
markdown==3.7
pydantic==2.10.6
shiny==1.4.0 
treelib==1.8.0
markitdown==0.1.2
python-dotenv==1.0.1